Add missing import renames
