+++ plugins/player_auth/test/plugin.js
@@ -7,7 +7,7 @@ const mock = require("../../../test/mock");
 
 const controller = require("../dist/plugin/controller");
 const instance = require("../dist/plugin/instance");
-const info = require("../dist/plugin/info").default;
+const info = require("../dist/plugin/index").plugin;
 const { FetchPlayerCodeRequest, SetVerifyCodeRequest } = require("../dist/plugin/messages");
 const lib = require("@clusterio/lib");
